chore: consolidate pattern types
