+++ src/core/manager.h
@@ -3,7 +3,7 @@
 
 #include "types.h"
 #include "memory/memory_tier_manager.hpp"
-#include "quantum/types.h"
+#include "core/types.h"
 #include <memory>
 #include <mutex>
 #include <string>
