+++ src/quantum/coherence_manager.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "quantum/types.h"
+#include "core/types.h"
 #include <glm/vec4.hpp>
 #include <array>
 #include <cstdint>
