+++ src/quantum/pattern_evolution_bridge.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "quantum/types.h"
+#include "core/types.h"
 #include <string>
 #include <vector>
 #include <memory>
