+++ src/quantum/processor.cpp
@@ -1,5 +1,5 @@
 #include "quantum/processor.h"
-#include "quantum/types.h"
+#include "core/types.h"
 #include "core/manager.h"
 #include "quantum/pattern_evolution_bridge.h"
 #include "quantum/quantum_processor_qfh.h"
