+++ src/quantum/processor.h
@@ -7,7 +7,8 @@
 #include "core/system_hooks.h"
 // Removed duplicate include of core/types.h
 #include "quantum/gpu_context.h"
-#include "quantum/data.hpp"
+#include "core/types.h"
+namespace sep { namespace quantum { using Pattern = ::sep::Pattern; } }
 #include "memory/types.h"
 #include <memory>
 #include <vector>
