+++ src/quantum/quantum_processor_qfh_common.cpp
@@ -1,6 +1,6 @@
 #include "compat/cuda.h"
 #include "compat/math_common.h"
-#include "quantum/types.h"
+#include "core/types.h"
 #include "quantum/qbsa_qfh.h"
 #include "quantum/quantum_processor.h"
 #include "quantum/quantum_processor_qfh.h"
