+++ src/sep_engine_wrapper.cpp
@@ -1,5 +1,6 @@
 #include "sep_engine_wrapper.h"
-#include "quantum/types.h"
+#include "core/types.h"
+namespace sep { namespace quantum { using Pattern = ::sep::Pattern; } }
 
 #include <iostream>
