+++ src/workbench/demos/digital_physics_demo.hpp
@@ -6,7 +6,7 @@
 #include "../../workbench_demo_adapter.hpp"
 #include "demo_manager.hpp"
 #include "imgui.h"
-#include "quantum/data.hpp"
+#include "core/types.h"
 
 namespace sep {
 namespace workbench {
